Fix my definition
As it turns out my definition of web applications is wrong in a sense.
I originally thought that a webapp was the HTML presentation in the
browser; and the PHP and database logic on the server. In a sense it is,
but my definition is describes a 1-tier architecture. A webapp is in
fact the entire client-server relationship. A webapp involves the entire
web. My definition was "an software application that users use via a web
browser." However it turns own that web application is a client/server
software application. So it is an application that involves a client
that gets data from a server. Now a client isn't just limited to
a browser: it can be a mobile app, a browser, or a native desktop
application.